requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
faker==20.1.0
bandit==1.7.5
safety==2.3.5

